Reproducibility and chain regularization
Sverre's email from 10/12/2009:

The important problem of reproducibility came up again, most likely
because of the parallel routine nbintp.f in the GPU version. It was
found that repeated runs do not give identical results after chain
regularization has been used. Further investigations revealed that
prediction of global chain variables does not always occur for the
same sequence in one block-step when using multiple threads. This
weakness is now avoided by making the following small changes:

1. chfirr.f:  CALL XCPRED is not needed (done at start of block-step).
              Secondly, CALL CHLIST only done for arg > 0.

2. xcpred.f:  Prediction of chain c.m. is not needed for
              argument KCASE=2.

3. intgrt.omp.f: CALL XCPRED(2) is included before
              NBINT/NBINTP if NCH > 0.

4. intgrt.f (standard NBODY6): Same as #3 for consistency
              (before label 50).

All four routines will be sent. Note there is no need to change anything
for use of NBODY6 without GPU but the code should be consistent.

Reproducibility was verified for long runs involving some 100 chain reg.
Another less elegant way would be to avoid using NBINTP on rare
occasions when NCH > 0 but this is not recommended.

Rainer's response from 10/12/2009:

the same problem has already been dealt with in NBODY6++ in Nov. 2003,
for those of you also occasionally using NBODY6++... I attach some code
fragment used in intgrt.F to get all the XC and UC outside of the
parallel block, before calling nbint.f . This was just for exactly the
same problem - reproducibility. This way avoids any changes in
chfirr.f and xcpred.f (no harm to call it again).

Frankly, when I now look back at my algorithm I still see a possible
bug (in the sense that reproducibility is in danger) if a chain is a
neighbour of an active particle, but not itself active (appears not
very likely to ne, but could happen...). I'll have to think about
this, preparing a test case can be _very_ cumbersome.
I think Sverre has now caught this second possibility with his call of
XCPRED(2) (KCASE=2 newly defined). That should do it as well for
intgrt.F.

PS: Oh, just for correctness of the history, Marc Hemsendorf had
detected the reproducibility problem in parallel blocks containing the
chain particle altready a few years earlier (2001) and cured it by a
global communication afterwards; what I sent you in the last mail was a
correction to take care of the problem _before_ a block.

Dear colleagues, sorry for the mails, but it turned out that one point
should be clarified for the record - the first to notice this error and
include a fix in NBODY6++ was Holger Baumgardt, way past at our ARI
times.

----------------(Code Fragment Nbody6++; INTGRT.F)---------------------
*       Resolve Chain if it is in block or in neighbour lists (RS Nov. 03).
      IF(NCH.GT.0) THEN
      ICHPR = 0
*       First check whether chain c.m. is in current block (ICHPR = 2).
          DO 46 L = 1,NXTLEN
              I = NXTLST(L)
              IF(I.EQ.ICH) ICHPR = 2
 46       CONTINUE
*       Second check whether neighbour lists contain chain c.m. (ICHPR = 1).
          IF (ICHPR.EQ.0) THEN
              DO 47 L = 1,NXTLEN
                  I = NXTLST(L)
                  NNB1 = LIST(1,I) + 1
*       Second check whether neighbour lists contain chain c.m. (ICHPR = 1).
                  DO 48 K = 2,NNB1
                      J = LIST(K,I)
                      IF (J.GT.ICH) GO TO 47
                      IF (J.EQ.ICH) ICHPR = 1
 48               CONTINUE
 47           CONTINUE
          END IF
          IF (ICHPR.GT.1) CALL CHLIST(ICH)
          IF (ICHPR.GT.0) CALL XCPRED(0)
      END IF
*