Wide binary perturber in stability test
From Sverre's email (04/21/10):

Dear NBODY6 users,

Christoph Olczak has drawn my attention to a rare bug in histab.f.
Thus in the case of a wider binary being the closest perturber in the
stability test, the switch fails to assign the new c.m. index used
for obtaining the inclination. This procedure is done correctly in
impact.f but here more care is needed to avoid changing the calling
argument.

histab.f:  to be replaced.

Best wishes,
Sverre