Collision ghost removal
From Sverre's email (04/26/10):

Dear NBODY6 and GPU users,

Here are three more bug fixes, with the first two for the standard code.

escape.f:

Nick reported problems due to a collision ghost not being removed.
Closer inspection showed that when using a large Plummer mass, the escape
condition EI > 0 was not satisfied.  It was also noticed that the Plummer
potential was in fact included a second time. This would affect the energy
check.