bug in bug report (histab.f;Wide binary perturber)
From Sverre's email (05/03/2010):

Dear NBODY6 users,

Yet again a bug fix requires repair. Thus in the recent new version
of histab.f we need to set JJ = J outside the IF test at line 76.
Somehow I was persuaded that the ELSE would take care of this but
there are two of them.

Sorry about the omission in histab.f.
Sverre