safety expression for convergent binding energy
From Sverre's email (05/03/2010):

ksint.f:  After watching a bad sequence of KS integration it seemed
desirable to introduce the safety expression for convergent binding
energy, this time only for significant perturbation.  This procedure is
already included in kspoly.f.  The extra number of step reductions was
extremely small in several realistic tests with primordial binaries so
this is really an insurance for rare situations.  The present problem
was seen in a simulation by Nick.