sweep.f, DTADJ and KS components in chain
From Sverre's email (05/24/2010):

Dear NBODY6 and GPU users,

I would like to remind you that routine sweep.f should not be called if
using output intervals DTADJ which are not commensurate with 1/2^n. It
is called at the output time so for example, DTADJ = 0.05 would lead to
a time-step near 1D-12 after an enforced KS regularization, and instead
of crashing might lead to the integration proceeding with tiny steps.
This problem can be overcome by including a condition such as
.AND.DMOD(TIME,DTK(10)).EQ.0.0D0 before CALL SWEEP. This would allow
some calls, like at TIME = 0.25 for DTADJ = 0.05.

Since this was written, Seungkuyng Oh has found  another problem in
sweep.f which also caused a crash. Namely a skip is missing if either
of the KS components are an existing chain c.m.