timing comparisons
From Sverre's  email (06/18/2010):

Dear GPU users,

I have been asked to prepare a report on the use of the Tesla GPU which
was supplied by NVIDIA via the University. Some of you may find it of
interest so I include it here together with a figure of timing
comparisons.  At the same time I include the correction of a minor
inconsistency.

lagr.f:  All TTOT appearing in IF tests (only) should be replaced by
TIME because TNEXT is reduced by the offset time with option #35.

Hot news: Keigo is arriving on Monday for his annual visit. We
have bought two GeForce GTX 470 cards which will replace the
Tesla. The idea is to implement the irregular force calculation
on the GPU and thereby reduce a significant bottleneck.

Best wishes, Sverre

PS. Since I am unable to send the figure using alias email
addresses I only include the LaTeX file here. Please request
the figure if you need it.